Switched from GITHUB_ENV to GITHUB_OUTPUT (step outputs)
I couldn’t find any official statement on which is the recommended approach.
But I did see this warning that says “set-output” is deprecated, switch to using environment files:
https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

Note that this doesn’t say that “steps.output” is deprecated; it says “set-output” is deprecated.
So, step outputs are still valid and we essentially have to choose between: GITHUB_ENV and GITHUB_OUTPUT

One argument in favor of GITHUB_OUTPUT is this:
- This comment mentions that Environment variables are only available in the same job.
- https://github.com/orgs/community/discussions/55294#discussioncomment-5884935

- Github_Output documentation mentions that job outputs available to all dependent downstream jobs.
- https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/passing-information-between-jobs#:~:text=Job%20outputs%20are%20available%20to%20all%20downstream%20jobs%20that%20depend%20on%20this%20job

- Right now we do have just one job in the dashboard build workflows.
- But in the future if we have more jobs, then GITHUB_OUTPUT could be more suited.